A test case setup structure that can be used to build a test case.
There is a setup for each type of element that makes up a case. Some additional typings have been introduced into a sequence, step, and action setups to allow action types to be extended and separated from each other.
A data scope is identified with a ScopeIdentifier, which is then going to be used considered when building a case.