Modifications to the build process of actions and steps.
Some changes to interfaces of action setups so that different types of actions have an ability to build action contexts of different types.
Changed the way how action setups are stored in an assertion step setup: an assertion step contains an underlying step setup of a corresponding type to which it delegates each action-related operation.